refactor out init into separate file for testing
